# Optimize PoW worker