--- a/go.mod
+++ b/go.mod
@@ -8,8 +8,9 @@ require (
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/stretchr/testify v1.6.1
 	golang.org/x/crypto v0.0.0-20200820211705-5c72a883971a
-	golang.org/x/sys v0.0.0-20200909081042-eff7692f9009 // indirect
 	google.golang.org/grpc v1.33.1
 	google.golang.org/protobuf v1.25.0
 	gopkg.in/h2non/gock.v1 v1.0.15
 )
+
+replace github.com/iotaledger/iota.go => github.com/wollac/iota.go v1.0.0-beta.9.0.20210211182213-ad429d349f5d