# Description of change

Optimizes the PoW worker by checking the number of trailing zeros directly from the batched Curl state. This improves the performance as the full hashes don't need to be computed.

This achieves the following speedup on amd64 over the current `chrysalis/phase2`:
```
name       old time/op  new time/op  delta
Score      14.5µs ± 4%  12.8µs ± 5%  -11.96%
Worker     3.69µs ± 3%  1.62µs ± 5%  -56.16%
```

## Type of change

- Enhancement (a non-breaking change which adds functionality)